+++ CHANGELOG.md
@@ -4,7 +4,14 @@
 
 ### Deprecations and compatibility notes
 
-- Minimum version of dependencies updated to spatialite 5.1.
+- For all spatial overlays, **sliver polygons** are now removed from the output by
+  default. Polygons are considered slivers if they are narrower than a certain
+  tolerance. By default, the tolerance is 0.001 CRS units if the CRS of the input layers
+  is a projected CRS, 1e-7 if it is a geographic CRS. More information + information how
+  to change this default tolerance can be found in the documentation of
+  [`options.set_sliver_tolerance`](https://geofileops.readthedocs.io/en/latest/api/geofileops.options.set_sliver_tolerance.html).
+- Minimum version of dependencies updated to GDAL 3.8, pyogrio 0.8, geopandas 0.13 and
+  spatialite 5.1.
 
 ### Improvements
 
@@ -12,6 +19,7 @@
 - Add `concat` function (#746, #747)
 - Add `join` function (#751)
 - Add `add_columns` function (#768)
+- Remove polygon slivers by default for all overlay operations (#776)
 - Add `zip_geofile` and `unzip_geofile` functions (#754, #743)
 - Add `write_mode="append_add_fields"` option to `copy_layer` (#750)
 - Add an `include_duplicates` parameter to `union`, `intersection` and `identity` (#757)
@@ -24,6 +32,9 @@
 - Improve performance of the subdividividing used in many operations (#730)
 - Improve performance of `dissolve` (#748)
 - Improve performance of two-layer operations using `nb_parallel=1` (#692)
+- Add support for the `"pyogrio-arrow"` gfo io-engine, and use it by default. This
+  increases performance for reading and writing dataframes as well as for geopandas
+  based operations like `dissolve`,... (#392)
 - Alternative query for `clip` + default `subdivide_coords` to 15000 (#450)
 - Improve and speed up attribute column type detection in two-layer operations (#758)
 - Ensure that the featurecount is properly cached in GPKG files, also for older GDAL